Update repository/downloads links to GitHub.
Note: Left BitBucket link for Windows executable.  I *could* create a release
for 0.7.0a just to upload the binary to GitHub, but I don't want that to show
up at the top when people are looking for latest releases.

If I ever get around to compiling a new Windows installer, then I'll attach
it with the latest release.